requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dnspython==1.16.0
dnspython3==1.15.0
idna==2.9
lxml==4.5.1
multitasking==0.0.9
numpy==1.18.4
pandas==1.0.3
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
yfinance==0.1.54
